feat(identity): implement declared proof claims handling in identity models and UI for enhanced user clarity
